Update and delete api have been created for the admin on all tables

+++ src/app/shopping-cart/shopping-cart.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-shopping-cart',
@@ -10,6 +11,7 @@ export class ShoppingCartComponent implements OnInit {
   productCart: Array<{_id: number, title: string, description: string, price: number, imagePath: string, qty: number}> = [];
   // price: number;
   totalPrice:number;
+  showMenu:boolean = false;
 
   constructor() { 
     
@@ -24,6 +26,9 @@ export class ShoppingCartComponent implements OnInit {
   ngOnInit() {
 
   }
+  menuShow(){
+    this.showMenu = true;
+  }
 
   createProductObj(){
 
@@ -37,32 +42,59 @@ export class ShoppingCartComponent implements OnInit {
     }  
   }
 
-  reduceByOne(product:any){
-    
-    console.log(product);
-    
-    for(var i=0; i< this.productCart.length; i++){
+  reduceByOne(product:any) {
+    if(confirm("Are you sure to reduce by one the "+product.title+" is good game")) {
+      for(var i=0; i< this.productCart.length; i++){
 
-      if(this.productCart[i]._id == product._id){
-        
-        console.log("Clicked reduceByOne");
-        this.productCart[i].price  /= this.productCart[i].qty;
-        this.totalPrice -= this.productCart[i].price;
-        this.productCart[i].qty -= 1;
-        this.productCart[i].price *= this.productCart[i].qty;
-      }
+        if(this.productCart[i]._id == product._id){
+          
+          console.log("Clicked reduceByOne");
+          this.productCart[i].price  /= this.productCart[i].qty;
+          this.totalPrice -= this.productCart[i].price;
+          this.productCart[i].qty -= 1;
+          this.productCart[i].price *= this.productCart[i].qty;
+        }
 
-      if(this.productCart[i].qty == 0){
-        
-        console.log(this.totalPrice);
-        this.productCart.splice(i, 1);    
-        localStorage.setItem("products", JSON.stringify(this.productCart));
-        var qtyP = this.productCart.length;
-        localStorage.setItem('qty', qtyP.toString());
+        if(this.productCart[i].qty == 0){
+          
+          console.log(this.totalPrice);
+          this.productCart.splice(i, 1);    
+          localStorage.setItem("products", JSON.stringify(this.productCart));
+          var qtyP = this.productCart.length;
+          localStorage.setItem('qty', qtyP.toString());
+        }
       }
-       
+      this.showMenu = false;
     }
+    this.showMenu = false;
   }
+
+  // reduceByOne(product:any){
+    
+  //   console.log(product);
+    
+    // for(var i=0; i< this.productCart.length; i++){
+
+    //   if(this.productCart[i]._id == product._id){
+        
+    //     console.log("Clicked reduceByOne");
+    //     this.productCart[i].price  /= this.productCart[i].qty;
+    //     this.totalPrice -= this.productCart[i].price;
+    //     this.productCart[i].qty -= 1;
+    //     this.productCart[i].price *= this.productCart[i].qty;
+    //   }
+
+    //   if(this.productCart[i].qty == 0){
+        
+    //     console.log(this.totalPrice);
+    //     this.productCart.splice(i, 1);    
+    //     localStorage.setItem("products", JSON.stringify(this.productCart));
+    //     var qtyP = this.productCart.length;
+    //     localStorage.setItem('qty', qtyP.toString());
+    //   }
+       
+    // }
+  // }
 
   removeAll(product:any){
